ppc/xive2: Introduce a XIVE2 core framework
The XIVE2 interrupt controller of the POWER10 processor follows the
same logic than on POWER9 but the HW interface has been largely
reviewed. The interrupt controller has a new register interface,
different BARs, extra VSDs, these will be described when we add the
device model for the baremetal machine.

The XIVE internal structures for the EAS, END, NVT have a different
layout which is a problem for the current core XIVE framework. To
avoid adding too much complexity in the XIVE models, a new XIVE2 core
framework is introduced. It duplicates the models which are closely
linked to the XIVE internal structures : Xive2Router and Xive2ENDSource
and reuses the XiveSource, XivePresenter, XiveTCTX models, as they are
more generic.

Signed-off-by: Cédric Le Goater <clg@kaod.org>